pyproject.toml
# Copyright (c) 2022-2023 Adam Karpierz
# Licensed under the zlib/libpng License
# https://opensource.org/licenses/Zlib

[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools>=67.8.0', 'wheel>=0.40.0', 'packaging>=23.1.0', 'tox>=3.28.0,<4']

[project]
name        = 'py-chocolatey'
version     = '0.1.0'
description = 'API for Chocolatey, the Package Manager for Windows.'
license = { text = 'zlib/libpng License ; https://opensource.org/licenses/Zlib' }
authors = [
    { name  = 'Adam Karpierz' },
    { email = 'adam@karpierz.net' },
]
maintainers = [
    { name  = 'Adam Karpierz' },
    { email = 'adam@karpierz.net' },
]
keywords = ['chocolatey', 'choco']
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: zlib/libpng License',
    'Operating System :: OS Independent',
    'Natural Language :: Polish',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: Implementation :: CPython',
    'Topic :: Software Development :: Libraries :: Python Modules',
]
requires-python = '>=3.7.0,<4.0.0'
dependencies = [
    # mandatory
    'setuptools>=67.8.0',
    'pkg_about>=1.0.8',
    # others
    'atpublic>=3.1.2',
    'nocasedict>=2.0.1',
    'regex>=2023.6.3',
]
dynamic = ['readme']
[project.optional-dependencies]
doc = [
    'docutils>=0.14,<0.19',
    'Sphinx>=5.3.0',
    'sphinx-tabs>=3.4.1',
    'sphinx-copybutton>=0.5.1',
    'sphinxcontrib-spelling>=7.7.0',
    'sphinx-lint>=0.6.7',
    'restructuredtext-lint>=1.4.0',
    'nbsphinx>=0.8.10',
]
test = [
    'deepdiff>=6.3.0',
    'rich>=13.6.0',
]
[project.urls]
Homepage      = 'https://pypi.org/project/py-chocolatey/'
Documentation = 'https://py-chocolatey.readthedocs.io/'
Download      = 'https://pypi.org/project/py-chocolatey/'
Source        = 'https://github.com/karpierz/chocolatey'
Issues        = 'https://github.com/karpierz/chocolatey/issues'

[project.scripts]
#'chocolatey' = 'chocolatey.__main__:main'

[project.gui-scripts]
#'chocolatey-gui' = 'chocolatey.gui:main'

#[project.entry-points.'chocolatey.magical']
#epoint = 'chocolatey:main_epoint'

[tool.setuptools.dynamic]
readme = { file = ['README.rst', 'CHANGES.rst'], content-type = 'text/x-rst; charset=UTF-8' }

[tool.setuptools]
license-files = ['LICENSE']
include-package-data = true
platforms = ['any']
zip-safe = true

[tool.setuptools.packages.find]
namespaces = false
where = ['src']

[tool.setuptools.package-dir]
'' = 'src'
#'chocolatey.tests' = 'tests'

[tool.setuptools.package-data]
chocolatey = [
]

[tool.coverage.run]
source = [
    'chocolatey',
    'tests',
]
omit = [
]
data_file = '.tox/coverage/.coverage'

[tool.coverage.report]
exclude_lines = [
    # Regexes
    # Have to re-enable the standard pragma
    'pragma: no cover',
    'if\s+self\.debug\s+:',
    'if\s+__debug__\s+:',
    'raise (AssertionError|NotImplementedError)',
    'if\s+(0|False)\s*:',
    '''if\s+__name__.*\s*==\s*'__main__'\s*:''',
    'if\s+__name__.*\s*==\s*"__main__"\s*:',
    '^\s*@unittest\.skip\(',
    '^\s*@unittest\.skipIf\(sys\.platform\.startswith\("win"\)',
]
omit = [
    'tests/run.py',
]
skip_covered = false

[tool.coverage.html]
directory = '.tox/coverage/.coverage.html'

[tool.flake8]
filename = ['*.py','*.pyx']
#include = ['tests']
#exclude = ['.tox','*.egg','.git','__pycache__','build','_build','docs/_build','dist']
max-line-length = 99
ignore = ['E126','E203','E221','E251','E302','E701','E702','E731',
          'E122','E127','E128','E222','E272','E241','E266','E226','E704',
          'D100','D101','D102','D103','D104','D400','D401','D202',
          'N806','N802','N803','N801',
          'I100','W503']
# (e.g. 'E4','W') default: 'E121','E123','126','226','E24','704'
#select =
#select = ['E','W','F','N','I']
output-file = '.tox/lint/flake8out.txt'
count = true
#show-pep8,
#show-source
#verbose
#quiet

#
# Configuration(s) for tox
#

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{37,38,39,310,311}, docs
minversion = 3.28.0,<4
isolated_build = True
skip_missing_interpreters = true
requires =
    {[base]setup_requires}
    tox-venv>=0.4.0
    tox-backticks>=0.3.0
    tox-tags>=0.2.0
    tox-run-before>=0.1.0

[base]
setup_requires =
    pip>=23.1.2
    setuptools>=67.8.0
    wheel>=0.40.0
packagesubdir = chocolatey

[testenv]
passenv = WINDIR
commands =
    {envpython} --version
    {envpython} -B -m tests {posargs}
extras = test
deps =
    {[base]setup_requires}

[testenv:prepare]
basepython = python3.7
skip_install = true
allowlist_externals =
    cmd
    .build
commands =
    cmd /C if exist .build.cmd .build.cmd

[testenv:coverage]
basepython = python3.7
commands =
    {envpython} -m coverage erase
    -{envpython} -B -m coverage run -m tests {posargs}
    {envpython} -m coverage report
    {envpython} -m coverage html
deps =
    {[testenv]deps}
    coverage>=7.2.7
    diff-cover>=7.6.0

[testenv:docs]
basepython = python3.7
commands =
    {envpython} -m sphinxlint --ignore .tox --ignore build --ignore dist
   #{envpython} -m sphinx.apidoc -f {envsitepackagesdir}/{[base]packagesubdir}
    {envpython} -c "from setuptools import setup ; setup()" -v build_sphinx -b html -E
   #{envpython} -c "from setuptools import setup ; setup()" -v build_sphinx -b linkcheck
    {envpython} -c "from setuptools import setup ; setup()" -v build_sphinx -b doctest
extras = doc
deps =

[testenv:build]
basepython = python3.7
setenv =
    PKG_PVER=`{envpython} -c "import platform ; print(''.join(platform.python_version_tuple()[:2]))" 2> nul`
    PKG_NAME=`{envpython} -c "import warnings ; warnings.filterwarnings('ignore') ;import setuptools ; setuptools._distutils.core._setup_stop_after='config' ; print(setuptools.setup().metadata.get_name(),     end='')" 2> nul`
    PKG_DIST=`{envpython} -c "import warnings ; warnings.filterwarnings('ignore') ;import setuptools ; setuptools._distutils.core._setup_stop_after='config' ; print(setuptools.setup().metadata.get_fullname(), end='')" 2> nul`
commands =
    {[testenv]commands}
    {[testenv:docs]commands}
    {envpython} -c "from setuptools import setup ; setup()" --quiet sdist --formats=zip
    {envpython} -c "from setuptools import setup ; setup()" --quiet bdist_wheel
    # check out for PyPi
    {envpython} -m twine check dist/*
extras =
    {[testenv]extras}
    {[testenv:docs]extras}
deps =
    {[testenv]deps}
    {[testenv:docs]deps}
    twine>=4.0.2

[testenv:publish]
basepython = python3.7
setenv =
    {[testenv:build]setenv}
commands =
    {[testenv:build]commands}
    # publish on PyPi
    {envpython} -m twine upload dist/*
extras =
    {[testenv:build]extras}
deps =
    {[testenv:build]deps}
    twine>=4.0.2

[testenv:lint]
basepython = python3.7
commands =
    #{envpython} -m flake8p {envsitepackagesdir}/{[base]packagesubdir}/
extras =
deps =
    {[testenv]deps}
    flake8>=5.0.4
    flake8-pyproject>=1.2.3
    flake8-docstrings>=1.7.0
    pep8-naming>=0.13.3
    flake8-builtins>=2.1.0
"""
